+++ src/node/graph.go
@@ -18,7 +18,7 @@ type Graph struct {
 
 func (g *Graph) GetBlocks() []poset.Block {
 	res := []poset.Block{}
- 	blockIdx := 0
+ 	blockIdx := int64(0)
 	store := g.Node.core.poset.Store
 
  	for blockIdx <= store.LastBlockIndex() {
@@ -55,7 +55,7 @@ func (g *Graph) GetParticipantEvents() map[string]map[string]poset.Event {
 
 		selfParent := fmt.Sprintf("Root%d", p.ID)
 
-		flagTable := make(map[string]int)
+		flagTable := make(map[string]int64)
 		flagTable[selfParent] = 1
 
 		// Create and save the first Event
@@ -83,7 +83,7 @@ func (g *Graph) GetParticipantEvents() map[string]map[string]poset.Event {
 func (g *Graph) GetRounds() []poset.RoundInfo {
 	res := []poset.RoundInfo{}
 
-	round := 0
+	round := int64(0)
 
 	store := g.Node.core.poset.Store
