+++ src/proxy/proxy.go
@@ -10,7 +10,7 @@ import (
 type AppProxy interface {
 	SubmitCh() chan []byte
 	CommitBlock(block poset.Block) ([]byte, error)
-	GetSnapshot(blockIndex int) ([]byte, error)
+	GetSnapshot(blockIndex int64) ([]byte, error)
 	Restore(snapshot []byte) error
 }
